separate styling for App, fix flexbox flaws in JokeDiv
