+++ src/index.html
@@ -1,13 +1,16 @@
 <!doctype html>
 <html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="preconnect" href="https://fonts.gstatic.com">
-<link href="https://fonts.googleapis.com/css2?family=Work+Sans:wght@400;500&display=swap" rel="stylesheet">
-    <title>React App</title>
-  </head>
-  <body style="min-height: 100vh; box-sizing: border-box; margin: 0">
-    <div id="root" style="min-height: 100vh; box-sizing: border-box; margin: 0"></div>
-  </body>
+
+<head>
+  <meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="preconnect" href="https://fonts.gstatic.com">
+  <link href="https://fonts.googleapis.com/css2?family=Work+Sans:wght@400;500&display=swap" rel="stylesheet">
+  <title>React Jokes App</title>
+</head>
+
+<body style="min-height: 100vh; box-sizing: border-box; margin: 0">
+  <div id="root" style="min-height: 100vh; box-sizing: border-box; margin: 0"></div>
+</body>
+
 </html>
